+++ app/src/main/java/com/eatssu/android/ui/main/LunchFragment.kt
@@ -12,19 +12,18 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.eatssu.android.adapter.*
-import com.eatssu.android.util.RetrofitImpl.retrofit
 import com.eatssu.android.data.enums.Restaurant
 import com.eatssu.android.data.enums.Time
+import com.eatssu.android.data.repository.MenuRepository
 import com.eatssu.android.data.service.MenuService
 import com.eatssu.android.databinding.FragmentLunchBinding
-import com.eatssu.android.data.repository.MenuRepository
 import com.eatssu.android.ui.info.*
+import com.eatssu.android.ui.main.calendar.CalendarViewModel
 import com.eatssu.android.ui.main.menu.FixedMenuAdapter
 import com.eatssu.android.ui.main.menu.MenuViewModel
+import com.eatssu.android.ui.main.menu.MenuViewModelFactory
 import com.eatssu.android.ui.main.menu.TodayMealAdapter
-import com.eatssu.android.ui.main.calendar.CalendarViewModel
-import com.eatssu.android.data.viewmodelFactory.MenuViewModelFactory
+import com.eatssu.android.util.RetrofitImpl.retrofit
 import dagger.hilt.android.AndroidEntryPoint
 import retrofit2.*
 import java.time.DayOfWeek
