+++ app/src/main/java/com/eatssu/android/ui/review/list/ReviewViewModelFactory.kt
@@ -1,18 +1,16 @@
-package com.eatssu.android.data.viewmodelFactory
+package com.eatssu.android.ui.review.list
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.eatssu.android.data.repository.ReviewRepository
 
-import com.eatssu.android.ui.review.write.UploadReviewViewModel
-
-class UploadReviewViewModelFactory(private val repository: ReviewRepository) :
+class ReviewViewModelFactory(private val repository: ReviewRepository) :
     ViewModelProvider.Factory {
 
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(UploadReviewViewModel::class.java)) {
+        if (modelClass.isAssignableFrom(ReviewViewModel::class.java)) {
             @Suppress("UNCHECKED_CAST")
-            return UploadReviewViewModel(repository) as T
+            return ReviewViewModel(repository) as T
         }
         throw IllegalArgumentException("Unknown ViewModel class")
     }
