+++ app/src/main/java/com/eatssu/android/ui/review/write/UploadReviewViewModelFactory.kt
@@ -1,18 +1,16 @@
-package com.eatssu.android.data.viewmodelFactory
+package com.eatssu.android.ui.review.write
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.eatssu.android.data.repository.ReviewRepository
 
-import com.eatssu.android.ui.review.list.ReviewViewModel
-
-class ReviewViewModelFactory(private val repository: ReviewRepository) :
+class UploadReviewViewModelFactory(private val repository: ReviewRepository) :
     ViewModelProvider.Factory {
 
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(ReviewViewModel::class.java)) {
+        if (modelClass.isAssignableFrom(UploadReviewViewModel::class.java)) {
             @Suppress("UNCHECKED_CAST")
-            return ReviewViewModel(repository) as T
+            return UploadReviewViewModel(repository) as T
         }
         throw IllegalArgumentException("Unknown ViewModel class")
     }
